Revert "refacotr(emitter): align file names with `microsoft/typespec` (#4704)"
This reverts commit 9b5d6e9ca06c4fa6cb1631317af0135b40721448.
