+++ src/TypeSpec.Extension/Emitter.Csharp/src/lib/clientModelBuilder.ts
@@ -30,28 +30,28 @@ import {
 import { getVersions } from "@typespec/versioning";
 import { $lib } from "../emitter.js";
 import { NetEmitterOptions, resolveOptions } from "../options.js";
-import { ClientKind } from "../type/client-kind.js";
-import { CodeModel } from "../type/code-model.js";
-import { InputClient } from "../type/input-client.js";
-import { InputConstant } from "../type/input-constant.js";
-import { InputOperation } from "../type/input-operation.js";
-import { InputOperationParameterKind } from "../type/input-operation-parameter-kind.js";
-import { InputParameter } from "../type/input-parameter.js";
+import { ClientKind } from "../type/clientKind.js";
+import { CodeModel } from "../type/codeModel.js";
+import { InputClient } from "../type/inputClient.js";
+import { InputConstant } from "../type/inputConstant.js";
+import { InputOperation } from "../type/inputOperation.js";
+import { InputOperationParameterKind } from "../type/inputOperationParameterKind.js";
+import { InputParameter } from "../type/inputParameter.js";
 import {
     InputEnumType,
     InputModelType,
     InputPrimitiveType
-} from "../type/input-type.js";
-import { InputPrimitiveTypeKind } from "../type/input-primitive-type-kind.js";
-import { RequestLocation } from "../type/request-location.js";
+} from "../type/inputType.js";
+import { InputPrimitiveTypeKind } from "../type/inputPrimitiveTypeKind.js";
+import { RequestLocation } from "../type/requestLocation.js";
 import { Usage } from "../type/usage.js";
 import { logger } from "./logger.js";
 import { getUsages, navigateModels } from "./model.js";
 import { loadOperation } from "./operation.js";
-import { processServiceAuthentication } from "./service-authentication.js";
+import { processServiceAuthentication } from "./serviceAuthentication.js";
 import { resolveServers } from "./typespecServer.js";
 import { createContentTypeOrAcceptParameter } from "./utils.js";
-import { InputTypeKind } from "../type/input-type-kind.js";
+import { InputTypeKind } from "../type/inputTypeKind.js";
 
 export function createModel(
     sdkContext: SdkContext<NetEmitterOptions>
