+++ src/TypeSpec.Extension/Emitter.Csharp/src/lib/operation.ts
@@ -26,13 +26,13 @@ import {
     HttpOperationResponse
 } from "@typespec/http";
 import { NetEmitterOptions } from "../options.js";
-import { BodyMediaType, typeToBodyMediaType } from "../type/body-media-type.js";
-import { collectionFormatToDelimMap } from "../type/collection-format.js";
-import { HttpResponseHeader } from "../type/http-response-header.js";
-import { InputConstant } from "../type/input-constant.js";
-import { InputOperation } from "../type/input-operation.js";
-import { InputOperationParameterKind } from "../type/input-operation-parameter-kind.js";
-import { InputParameter } from "../type/input-parameter.js";
+import { BodyMediaType, typeToBodyMediaType } from "../type/bodyMediaType.js";
+import { collectionFormatToDelimMap } from "../type/collectionFormat.js";
+import { HttpResponseHeader } from "../type/httpResponseHeader.js";
+import { InputConstant } from "../type/inputConstant.js";
+import { InputOperation } from "../type/inputOperation.js";
+import { InputOperationParameterKind } from "../type/inputOperationParameterKind.js";
+import { InputParameter } from "../type/inputParameter.js";
 import {
     InputEnumType,
     InputListType,
@@ -42,19 +42,19 @@ import {
     isInputLiteralType,
     isInputModelType,
     isInputUnionType
-} from "../type/input-type.js";
-import { convertLroFinalStateVia } from "../type/operation-final-state-via.js";
-import { OperationLongRunning } from "../type/operation-long-running.js";
-import { OperationPaging } from "../type/operation-paging.js";
-import { OperationResponse } from "../type/operation-response.js";
+} from "../type/inputType.js";
+import { convertLroFinalStateVia } from "../type/operationFinalStateVia.js";
+import { OperationLongRunning } from "../type/operationLongRunning.js";
+import { OperationPaging } from "../type/operationPaging.js";
+import { OperationResponse } from "../type/operationResponse.js";
 import {
     RequestLocation,
     requestLocationMap
-} from "../type/request-location.js";
+} from "../type/requestLocation.js";
 import {
     parseHttpRequestMethod,
     RequestMethod
-} from "../type/request-method.js";
+} from "../type/requestMethod.js";
 import { getExternalDocs, getOperationId, hasDecorator } from "./decorators.js";
 import { logger } from "./logger.js";
 import {
@@ -68,7 +68,7 @@ import {
     getTypeName
 } from "./utils.js";
 import { Usage } from "../type/usage.js";
-import { InputTypeKind } from "../type/input-type-kind.js";
+import { InputTypeKind } from "../type/inputTypeKind.js";
 
 export function loadOperation(
     sdkContext: SdkContext<NetEmitterOptions>,
