+++ src/TypeSpec.Extension/Emitter.Csharp/src/lib/typespecServer.ts
@@ -3,20 +3,20 @@
 
 import { getDoc } from "@typespec/compiler";
 import { HttpServer } from "@typespec/http";
-import { InputConstant } from "../type/input-constant.js";
-import { InputOperationParameterKind } from "../type/input-operation-parameter-kind.js";
-import { InputParameter } from "../type/input-parameter.js";
+import { InputConstant } from "../type/inputConstant.js";
+import { InputOperationParameterKind } from "../type/inputOperationParameterKind.js";
+import { InputParameter } from "../type/inputParameter.js";
 import {
     InputEnumType,
     InputModelType,
     InputPrimitiveType,
     InputType
-} from "../type/input-type.js";
-import { InputPrimitiveTypeKind } from "../type/input-primitive-type-kind.js";
-import { RequestLocation } from "../type/request-location.js";
+} from "../type/inputType.js";
+import { InputPrimitiveTypeKind } from "../type/inputPrimitiveTypeKind.js";
+import { RequestLocation } from "../type/requestLocation.js";
 import { getDefaultValue, getInputType } from "./model.js";
 import { SdkContext } from "@azure-tools/typespec-client-generator-core";
-import { InputTypeKind } from "../type/input-type-kind.js";
+import { InputTypeKind } from "../type/inputTypeKind.js";
 import { NetEmitterOptions } from "../options.js";
 
 export interface TypeSpecServer {
