+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/clientKind.ts
