+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/collectionFormat.ts
