+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/externalDocs.ts
