+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/httpResponseHeader.ts
@@ -1,15 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputType } from "./input-type.js";
+import { InputType } from "./inputType.js";
 
-export interface InputModelProperty {
+export interface HttpResponseHeader {
     Name: string;
-    SerializedName: string;
+    NameInResponse: string;
     Description: string;
     Type: InputType;
-    IsRequired: boolean;
-    IsReadOnly: boolean;
-    IsDiscriminator?: boolean;
-    FlattenedNames?: string[];
 }
