+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputApiKeyAuth.ts
