+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputClient.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputOperation } from "./input-operation.js";
-import { InputParameter } from "./input-parameter.js";
+import { InputOperation } from "./inputOperation.js";
+import { InputParameter } from "./inputParameter.js";
 import { Protocols } from "./protocols.js";
 
 export interface InputClient {
