+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputEnumTypeValue.ts
