+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputIntrinsicTypeKind.ts
