+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputModelProperty.ts
@@ -1,11 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputType } from "./input-type.js";
+import { InputType } from "./inputType.js";
 
-export interface HttpResponseHeader {
+export interface InputModelProperty {
     Name: string;
-    NameInResponse: string;
+    SerializedName: string;
     Description: string;
     Type: InputType;
+    IsRequired: boolean;
+    IsReadOnly: boolean;
+    IsDiscriminator?: boolean;
+    FlattenedNames?: string[];
 }
