+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputOAuth2Auth.ts
