+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputOperationParameterKind.ts
