+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputParameter.ts
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputConstant } from "./input-constant.js";
-import { InputOperationParameterKind } from "./input-operation-parameter-kind.js";
-import { InputType } from "./input-type.js";
-import { RequestLocation } from "./request-location.js";
+import { InputConstant } from "./inputConstant.js";
+import { InputOperationParameterKind } from "./inputOperationParameterKind.js";
+import { InputType } from "./inputType.js";
+import { RequestLocation } from "./requestLocation.js";
 
 //TODO: Define VirtualParameter for HLC
 export interface VirtualParameter {}
