+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputPrimitiveTypeKind.ts
