+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputTypeKind.ts
