+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputTypeSerializationFormat.ts
