+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/inputTypeValue.ts
