+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/literalTypeContext.ts
