+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/operationLongRunning.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { OperationFinalStateVia } from "./operation-final-state-via.js";
-import { OperationResponse } from "./operation-response.js";
+import { OperationFinalStateVia } from "./operationFinalStateVia.js";
+import { OperationResponse } from "./operationResponse.js";
 
 export interface OperationLongRunning {
     FinalStateVia: OperationFinalStateVia;
