+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/operationPaging.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { InputOperation } from "./input-operation.js";
+import { InputOperation } from "./inputOperation.js";
 
 export interface OperationPaging {
     NextLinkName?: string;
