+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/requestLocation.ts
