+++ src/TypeSpec.Extension/Emitter.Csharp/src/type/validationType.ts
