+++ src/TypeSpec.Extension/Emitter.Csharp/test/Unit/operationFinalStateVia.test.ts
@@ -4,7 +4,7 @@ import { FinalStateValue } from "@azure-tools/typespec-azure-core";
 import {
     OperationFinalStateVia,
     convertLroFinalStateVia
-} from "../../src/type/operation-final-state-via.js";
+} from "../../src/type/operationFinalStateVia.js";
 import assert from "assert";
 import { describe } from "mocha";
