+++ src/TypeSpec.Extension/Emitter.Csharp/test/Unit/scalar.test.ts
@@ -1,15 +1,15 @@
 import { TestHost } from "@typespec/compiler/testing";
 import assert, { deepStrictEqual } from "assert";
 import isEqual from "lodash.isequal";
-import { createModel } from "../../src/lib/client-model-builder.js";
-import { InputPrimitiveTypeKind } from "../../src/type/input-primitive-type-kind.js";
-import { InputTypeKind } from "../../src/type/input-type-kind.js";
+import { createModel } from "../../src/lib/clientModelBuilder.js";
+import { InputPrimitiveTypeKind } from "../../src/type/inputPrimitiveTypeKind.js";
+import { InputTypeKind } from "../../src/type/inputTypeKind.js";
 import {
     createEmitterContext,
     createEmitterTestHost,
     createNetSdkContext,
     typeSpecCompile
-} from "./utils/test-util.js";
+} from "./utils/TestUtil.js";
 
 describe("Test GetInputType for scalar", () => {
     let runner: TestHost;
